% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webmasters_objects.R
\name{ApiDataRow}
\alias{ApiDataRow}
\title{Search Console API Objects 
View Google Search Console data for your verified sites.}
\usage{
ApiDataRow(clicks = NULL, ctr = NULL, impressions = NULL, keys = NULL,
  position = NULL)
}
\arguments{
\item{clicks}{No description}

\item{ctr}{No description}

\item{impressions}{No description}

\item{keys}{No description}

\item{position}{No description}
}
\value{
ApiDataRow object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 23:50:42
filename: /Users/mark/dev/R/autoGoogleAPI/googlewebmastersv3.auto/R/webmasters_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
ApiDataRow Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

