% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webmasters_objects.R
\name{ApiDimensionFilter}
\alias{ApiDimensionFilter}
\title{ApiDimensionFilter Object}
\usage{
ApiDimensionFilter(dimension = NULL, expression = NULL, operator = NULL)
}
\arguments{
\item{dimension}{No description}

\item{expression}{No description}

\item{operator}{No description}
}
\value{
ApiDimensionFilter object
}
\description{
ApiDimensionFilter Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
