% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webmasters_objects.R
\name{ApiDimensionFilterGroup}
\alias{ApiDimensionFilterGroup}
\title{ApiDimensionFilterGroup Object}
\usage{
ApiDimensionFilterGroup(filters = NULL, groupType = NULL)
}
\arguments{
\item{filters}{No description}

\item{groupType}{No description}
}
\value{
ApiDimensionFilterGroup object
}
\description{
ApiDimensionFilterGroup Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

