% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webmasters_functions.R
\name{searchanalytics.query}
\alias{searchanalytics.query}
\title{Query your data with filters and parameters that you define. Returns zero or more rows grouped by the row keys that you define. You must define a date range of one or more days.When date is one of the group by values, any days without data are omitted from the result list. If you need to know which days have data, issue a broad date range query grouped by date for any metric, and see which day rows are returned.}
\usage{
searchanalytics.query(SearchAnalyticsQueryRequest, siteUrl)
}
\arguments{
\item{SearchAnalyticsQueryRequest}{The \link{SearchAnalyticsQueryRequest} object to pass to this method}

\item{siteUrl}{The site's URL, including protocol}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/webmasters
\item https://www.googleapis.com/auth/webmasters.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/webmasters, https://www.googleapis.com/auth/webmasters.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/webmaster-tools/}{Google Documentation}

Other SearchAnalyticsQueryRequest functions: \code{\link{SearchAnalyticsQueryRequest}}
}
