% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webmasters_objects.R
\name{SearchAnalyticsQueryRequest}
\alias{SearchAnalyticsQueryRequest}
\title{SearchAnalyticsQueryRequest Object}
\usage{
SearchAnalyticsQueryRequest(aggregationType = NULL,
  dimensionFilterGroups = NULL, dimensions = NULL, endDate = NULL,
  rowLimit = NULL, searchType = NULL, startDate = NULL, startRow = NULL)
}
\arguments{
\item{aggregationType}{[Optional; Default is 'auto'] How data is aggregated}

\item{dimensionFilterGroups}{[Optional] Zero or more filters to apply to the dimension grouping values; for example, 'query contains 'buy'' to see only data where the query string contains the substring 'buy' (not case-sensitive)}

\item{dimensions}{[Optional] Zero or more dimensions to group results by}

\item{endDate}{[Required] End date of the requested date range, in YYYY-MM-DD format, in PST (UTC - 8:00)}

\item{rowLimit}{[Optional; Default is 1000] The maximum number of rows to return}

\item{searchType}{[Optional; Default is 'web'] The search type to filter for}

\item{startDate}{[Required] Start date of the requested date range, in YYYY-MM-DD format, in PST time (UTC - 8:00)}

\item{startRow}{[Optional; Default is 0] Zero-based index of the first row in the response}
}
\value{
SearchAnalyticsQueryRequest object
}
\description{
SearchAnalyticsQueryRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other SearchAnalyticsQueryRequest functions: \code{\link{searchanalytics.query}}
}

