% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webmasters_objects.R
\name{SearchAnalyticsQueryResponse}
\alias{SearchAnalyticsQueryResponse}
\title{SearchAnalyticsQueryResponse Object}
\usage{
SearchAnalyticsQueryResponse(responseAggregationType = NULL, rows = NULL)
}
\arguments{
\item{responseAggregationType}{How the results were aggregated}

\item{rows}{A list of rows grouped by the key values in the order given in the query}
}
\value{
SearchAnalyticsQueryResponse object
}
\description{
SearchAnalyticsQueryResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A list of rows, one per result, grouped by key. Metrics in each row are aggregated for all data grouped by that key either by page or property, as specified by the aggregation type parameter.
}
