% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webmasters_functions.R
\name{sitemaps.delete}
\alias{sitemaps.delete}
\title{Deletes a sitemap from this site.}
\usage{
sitemaps.delete(siteUrl, feedpath)
}
\arguments{
\item{siteUrl}{The site's URL, including protocol}

\item{feedpath}{The URL of the actual sitemap}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/webmasters
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/webmasters)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/webmaster-tools/}{Google Documentation}
}

