% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webmasters_functions.R
\name{sitemaps.get}
\alias{sitemaps.get}
\title{Retrieves information about a specific sitemap.}
\usage{
sitemaps.get(siteUrl, feedpath)
}
\arguments{
\item{siteUrl}{The site's URL, including protocol}

\item{feedpath}{The URL of the actual sitemap}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/webmasters
\item https://www.googleapis.com/auth/webmasters.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/webmasters, https://www.googleapis.com/auth/webmasters.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/webmaster-tools/}{Google Documentation}
}

