% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webmasters_objects.R
\name{SitemapsListResponse}
\alias{SitemapsListResponse}
\title{SitemapsListResponse Object}
\usage{
SitemapsListResponse(sitemap = NULL)
}
\arguments{
\item{sitemap}{Contains detailed information about a specific URL submitted as a sitemap}
}
\value{
SitemapsListResponse object
}
\description{
SitemapsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
List of sitemaps.
}

