% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webmasters_functions.R
\name{sites.add}
\alias{sites.add}
\title{Adds a site to the set of the user's sites in Search Console.}
\usage{
sites.add(siteUrl)
}
\arguments{
\item{siteUrl}{The URL of the site to add}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/webmasters
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/webmasters)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/webmaster-tools/}{Google Documentation}
}

