% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webmasters_functions.R
\name{sites.delete}
\alias{sites.delete}
\title{Removes a site from the set of the user's Search Console sites.}
\usage{
sites.delete(siteUrl)
}
\arguments{
\item{siteUrl}{The URI of the property as defined in Search Console}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/webmasters
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/webmasters)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/webmaster-tools/}{Google Documentation}
}

