% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webmasters_functions.R
\name{sites.list}
\alias{sites.list}
\title{Lists the user's Search Console sites.}
\usage{
sites.list()
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/webmasters
\item https://www.googleapis.com/auth/webmasters.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/webmasters, https://www.googleapis.com/auth/webmasters.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details. 

#' @importFrom googleAuthR gar_api_generator
}
\seealso{
\href{https://developers.google.com/webmaster-tools/}{Google Documentation}
}

