% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webmasters_objects.R
\name{SitesListResponse}
\alias{SitesListResponse}
\title{SitesListResponse Object}
\usage{
SitesListResponse(siteEntry = NULL)
}
\arguments{
\item{siteEntry}{Contains permission level information about a Search Console site}
}
\value{
SitesListResponse object
}
\description{
SitesListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
List of sites with access level information.
}

