% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webmasters_objects.R
\name{UrlCrawlErrorCount}
\alias{UrlCrawlErrorCount}
\title{UrlCrawlErrorCount Object}
\usage{
UrlCrawlErrorCount(count = NULL, timestamp = NULL)
}
\arguments{
\item{count}{The error count at the given timestamp}

\item{timestamp}{The date and time when the crawl attempt took place, in RFC 3339 format}
}
\value{
UrlCrawlErrorCount object
}
\description{
UrlCrawlErrorCount Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An entry in a URL crawl errors time series.
}
