% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webmasters_objects.R
\name{UrlCrawlErrorCountsPerType}
\alias{UrlCrawlErrorCountsPerType}
\title{UrlCrawlErrorCountsPerType Object}
\usage{
UrlCrawlErrorCountsPerType(category = NULL, entries = NULL,
  platform = NULL)
}
\arguments{
\item{category}{The crawl error type}

\item{entries}{The error count entries time series}

\item{platform}{The general type of Googlebot that made the request (see list of Googlebot user-agents for the user-agents used)}
}
\value{
UrlCrawlErrorCountsPerType object
}
\description{
UrlCrawlErrorCountsPerType Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Number of errors per day for a specific error type (defined by platform and category).
}

