% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webmasters_functions.R
\name{urlcrawlerrorscounts.query}
\alias{urlcrawlerrorscounts.query}
\title{Retrieves a time series of the number of URL crawl errors per error category and platform.}
\usage{
urlcrawlerrorscounts.query(siteUrl, category = NULL,
  latestCountsOnly = NULL, platform = NULL)
}
\arguments{
\item{siteUrl}{The site's URL, including protocol}

\item{category}{The crawl error category}

\item{latestCountsOnly}{If true, returns only the latest crawl error counts}

\item{platform}{The user agent type (platform) that made the request}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/webmasters
\item https://www.googleapis.com/auth/webmasters.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/webmasters, https://www.googleapis.com/auth/webmasters.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/webmaster-tools/}{Google Documentation}
}
