% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webmasters_objects.R
\name{UrlCrawlErrorsCountsQueryResponse}
\alias{UrlCrawlErrorsCountsQueryResponse}
\title{UrlCrawlErrorsCountsQueryResponse Object}
\usage{
UrlCrawlErrorsCountsQueryResponse(countPerTypes = NULL)
}
\arguments{
\item{countPerTypes}{The time series of the number of URL crawl errors per error category and platform}
}
\value{
UrlCrawlErrorsCountsQueryResponse object
}
\description{
UrlCrawlErrorsCountsQueryResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A time series of the number of URL crawl errors per error category and platform.
}

