% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webmasters_objects.R
\name{UrlCrawlErrorsSample}
\alias{UrlCrawlErrorsSample}
\title{UrlCrawlErrorsSample Object}
\usage{
UrlCrawlErrorsSample(first_detected = NULL, last_crawled = NULL,
  pageUrl = NULL, responseCode = NULL, urlDetails = NULL)
}
\arguments{
\item{first_detected}{The time the error was first detected, in RFC 3339 format}

\item{last_crawled}{The time when the URL was last crawled, in RFC 3339 format}

\item{pageUrl}{The URL of an error, relative to the site}

\item{responseCode}{The HTTP response code, if any}

\item{urlDetails}{Additional details about the URL, set only when calling get()}
}
\value{
UrlCrawlErrorsSample object
}
\description{
UrlCrawlErrorsSample Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains information about specific crawl errors.
}
