% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webmasters_functions.R
\name{urlcrawlerrorssamples.get}
\alias{urlcrawlerrorssamples.get}
\title{Retrieves details about crawl errors for a site's sample URL.}
\usage{
urlcrawlerrorssamples.get(siteUrl, url, category, platform)
}
\arguments{
\item{siteUrl}{The site's URL, including protocol}

\item{url}{The relative path (without the site) of the sample URL}

\item{category}{The crawl error category}

\item{platform}{The user agent type (platform) that made the request}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/webmasters
\item https://www.googleapis.com/auth/webmasters.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/webmasters, https://www.googleapis.com/auth/webmasters.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/webmaster-tools/}{Google Documentation}
}
