% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webmasters_objects.R
\name{UrlCrawlErrorsSamplesListResponse}
\alias{UrlCrawlErrorsSamplesListResponse}
\title{UrlCrawlErrorsSamplesListResponse Object}
\usage{
UrlCrawlErrorsSamplesListResponse(urlCrawlErrorSample = NULL)
}
\arguments{
\item{urlCrawlErrorSample}{Information about the sample URL and its crawl error}
}
\value{
UrlCrawlErrorsSamplesListResponse object
}
\description{
UrlCrawlErrorsSamplesListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
List of crawl error samples.
}

