% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webmasters_objects.R
\name{UrlSampleDetails}
\alias{UrlSampleDetails}
\title{UrlSampleDetails Object}
\usage{
UrlSampleDetails(containingSitemaps = NULL, linkedFromUrls = NULL)
}
\arguments{
\item{containingSitemaps}{List of sitemaps pointing at this URL}

\item{linkedFromUrls}{A sample set of URLs linking to this URL}
}
\value{
UrlSampleDetails object
}
\description{
UrlSampleDetails Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Additional details about the URL, set only when calling get().
}

