% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webmasters_functions.R
\docType{package}
\name{webmasters_googleAuthR}
\alias{webmasters_googleAuthR}
\alias{webmasters_googleAuthR-package}
\title{Search Console API
View Google Search Console data for your verified sites.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-03 23:50:42
filename: /Users/mark/dev/R/autoGoogleAPI/googlewebmastersv3.auto/R/webmasters_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/webmasters
\item https://www.googleapis.com/auth/webmasters.readonly
}
}

