% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webmasters_objects.R
\name{WmxSite}
\alias{WmxSite}
\title{WmxSite Object}
\usage{
WmxSite(permissionLevel = NULL, siteUrl = NULL)
}
\arguments{
\item{permissionLevel}{The user's permission level for the site}

\item{siteUrl}{The URL of the site}
}
\value{
WmxSite object
}
\description{
WmxSite Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains permission level information about a Search Console site. For more information, see  Permissions in Search Console.
}

