% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webmasters_objects.R
\name{WmxSitemap}
\alias{WmxSitemap}
\title{WmxSitemap Object}
\usage{
WmxSitemap(contents = NULL, errors = NULL, isPending = NULL,
  isSitemapsIndex = NULL, lastDownloaded = NULL, lastSubmitted = NULL,
  path = NULL, type = NULL, warnings = NULL)
}
\arguments{
\item{contents}{The various content types in the sitemap}

\item{errors}{Number of errors in the sitemap}

\item{isPending}{If true, the sitemap has not been processed}

\item{isSitemapsIndex}{If true, the sitemap is a collection of sitemaps}

\item{lastDownloaded}{Date & time in which this sitemap was last downloaded}

\item{lastSubmitted}{Date & time in which this sitemap was submitted}

\item{path}{The url of the sitemap}

\item{type}{The type of the sitemap}

\item{warnings}{Number of warnings for the sitemap}
}
\value{
WmxSitemap object
}
\description{
WmxSitemap Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains detailed information about a specific URL submitted as a sitemap.
}
