% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webmasters_objects.R
\name{WmxSitemapContent}
\alias{WmxSitemapContent}
\title{WmxSitemapContent Object}
\usage{
WmxSitemapContent(indexed = NULL, submitted = NULL, type = NULL)
}
\arguments{
\item{indexed}{The number of URLs from the sitemap that were indexed (of the content type)}

\item{submitted}{The number of URLs in the sitemap (of the content type)}

\item{type}{The specific type of content in this sitemap}
}
\value{
WmxSitemapContent object
}
\description{
WmxSitemapContent Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Information about the various content types in the sitemap.
}
